- correction de l'affichage de la liste du champ co-adhérent - afficher le champ montant_total pour tous les modes de paiement et rendre le champ éditable - rendre champ telephone optionnel - ajouter une fausse adresse pour permettre d'enregistrer les adherent sans email (tout en maintenant le champ obligatoire)

+++ plugins/auto/amap_extra/amap_extra_fonctions.php
@@ -42,7 +42,7 @@ function action_supprimer_adherent($arg=null) {
             'id_adherent=' . sql_quote($arg)
         );
         $error = 0;
-		if ($mailinglist != '') {
+		if ($ancien_mail != lire_config('amap_extra/ovh_fakemail') and $mailinglist != '') {
 			try {
 				$ovh->delete('/email/domain/' . lire_config('amap_extra/ovh_domain') . '/mailingList/' . $mailinglist . '/subscriber/' . $ancien_mail);
 			} catch (GuzzleHttp\Exception\ClientException $e) {
